replace some search functions to mixin
